crud api

+++ Clothes_SpringBoot/src/main/java/edu/t3h/clothes/repository/ProducerReponsiroty.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProducerReponsiroty extends JpaRepository<ProducerEntity,Integer> {
+public interface ProducerReponsiroty extends JpaRepository<ProducerEntity,Long> {
     @Query("SELECT c FROM ProducerEntity c WHERE c.name LIKE %?1%")
     List<ProducerEntity> searchProducer(String keyword);
 }
